requirements.txt
pydantic~=2.11.4
SQLAlchemy~=2.0.40
requests~=2.32.3
fastapi~=0.115.9
python-jose~=3.4.0
python-dotenv~=1.1.0
numpy~=1.26.4
sentence-transformers~=4.1.0
torch~=2.7.0
unsloth~=2025.4.7
transformers~=4.51.3
peft~=0.15.2